+++ index.html
@@ -7,27 +7,9 @@
     <meta name="theme-color" content="#ffffff" media="(prefers-color-scheme: light)">
     <meta name="theme-color" content="#1f2937" media="(prefers-color-scheme: dark)">
     <title>Isuru Madusanka Rodrigo - Web Developer</title>
-    <link rel="icon" href="/assets/images/I.png" type="image/png">
-    <link href="css/style.css" rel="stylesheet">
+    <link rel="icon" href="assets/images/I.png" type="image/png">
+    <link href="/css/style.css" rel="stylesheet">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
-    <!-- Tailwind configuration -->
-    <script>
-        tailwind.config = {
-            darkMode: 'class',
-            theme: {
-                extend: {
-                    colors: {
-                        primary: '#3b82f6',
-                        secondary: '#06b6d4',
-                        accent: '#facc15',
-                        background: '#f9fafb',
-                        darkBackground: '#0f172a',
-                        softDark: '#1e293b'
-                    }
-                }
-            }
-        }
-    </script>
 </head>
 <body class="bg-gray-50 dark:bg-gray-900 transition-colors duration-300">
     <!-- Skip to main content -->
@@ -250,7 +232,7 @@
                             <div class="flex-grow">
                                 <h3 class="text-2xl font-bold text-gray-900 dark:text-white mb-2">Bachelor of Technology</h3>
                                 <h4 class="text-lg text-blue-600 dark:text-blue-400 mb-2">Rajarata University of Sri Lanka</h4>
-                                <p class="text-gray-600 dark:text-gray-300 mb-2">2020 – 2024</p>
+                                <p class="text-gray-600 dark:text-gray-300 mb-2">2020 – Present</p>
                                 <p class="text-gray-700 dark:text-gray-200 mb-4">ICT | Software Engineering | Blockchain | IoT 🧠💻</p>
                                 <div class="space-y-2">
                                     <p class="text-gray-600 dark:text-gray-300">• Final-year undergraduate specializing in software project development</p>
@@ -273,8 +255,9 @@
                                 <h3 class="text-2xl font-bold text-gray-900 dark:text-white mb-2">Diploma in Information Technology</h3>
                                 <h4 class="text-lg text-blue-600 dark:text-blue-400 mb-2">NIBM Kandy</h4>
                                 <p class="text-gray-600 dark:text-gray-300 mb-2">2018 – 2020</p>
-                                <p class="text-gray-700 dark:text-gray-200 mb-4">Software Engineering & Programming Fundamentals 👨‍💻📘</p>
+                                <p class="text-gray-700 dark:text-gray-200 mb-4">Diploma Software Engineering 👨‍💻📘</p>
                                 <div class="space-y-2">
+                                    <p class="text-gray-600 dark:text-gray-300">• GPA: 3.65 / 4.0</p>
                                     <p class="text-gray-600 dark:text-gray-300">• Mastered core programming (C, Python, Java)</p>
                                     <p class="text-gray-600 dark:text-gray-300">• Built first real-world projects</p>
                                     <p class="text-gray-600 dark:text-gray-300">• Developed interest in open source and IoT</p>
@@ -297,7 +280,7 @@
                                 <p class="text-gray-600 dark:text-gray-300 mb-2">2019</p>
                                 <p class="text-gray-700 dark:text-gray-200 mb-4">Technology Stream 🖥️⚙️</p>
                                 <div class="space-y-2">
-                                    <p class="text-gray-600 dark:text-gray-300">• Results: B2 | C1</p>
+                                    <p class="text-gray-600 dark:text-gray-300">• Results: B1 | C2</p>
                                     <p class="text-gray-600 dark:text-gray-300">• Strengthened analytical and technical focus in ICT</p>
                                 </div>
                             </div>
@@ -316,9 +299,8 @@
                                 <h3 class="text-2xl font-bold text-gray-900 dark:text-white mb-2">GCE Ordinary Level</h4>
                                 <h4 class="text-lg text-blue-600 dark:text-blue-400 mb-2">St. Thomas' College, Matale</h4>
                                 <p class="text-gray-600 dark:text-gray-300 mb-2">2016</p>
-                                <p class="text-gray-700 dark:text-gray-200 mb-4">Technology Stream 🖥️⚙️</p>
                                 <div class="space-y-2">
-                                    <p class="text-gray-600 dark:text-gray-300">• Results: B1 | C2</p>
+                                    <p class="text-gray-600 dark:text-gray-300">• Results: A5 | B2 | C2</p>
                                     <p class="text-gray-600 dark:text-gray-300">• First exposure to computing and structured problem solving</p>
                                 </div>
                             </div>
@@ -387,7 +369,7 @@
                             title="Close case study">
                         <i class="fas fa-times text-xl" aria-hidden="true"></i>
                     </button>
-                    <div class="p-8">
+                    <div class="p-8"></div>
                         <div id="case-study-content" class="space-y-8">
                             <h2 class="text-3xl font-bold text-gray-900 dark:text-white"></h2>
                             
@@ -591,10 +573,31 @@
     <!-- Scripts -->
     <script src="https://cdn.jsdelivr.net/npm/@emailjs/browser@3/dist/email.min.js"></script>
     <script src="https://cdn.tailwindcss.com" defer></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/gsap.min.js" defer></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/ScrollTrigger.min.js" defer></script>
-    <script src="https://unpkg.com/@lottiefiles/lottie-player@latest/dist/lottie-player.js" defer></script>
-    <script src="js/main.js" defer></script>
-    <script src="js/projects.js" defer></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/gsap.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/ScrollTrigger.min.js"></script>
+    <script src="https://unpkg.com/@lottiefiles/lottie-player@latest/dist/lottie-player.js"></script>
+    <!-- Tailwind configuration -->
+    <script>
+        document.addEventListener('DOMContentLoaded', () => {
+            if (typeof tailwind !== 'undefined') {
+                tailwind.config = {
+                    darkMode: 'class',
+                    theme: {
+                        extend: {
+                            colors: {
+                                primary: '#3b82f6',
+                                secondary: '#06b6d4',
+                                accent: '#facc15',
+                                background: '#f9fafb',
+                                darkBackground: '#0f172a',
+                                softDark: '#1e293b'
+                            }
+                        }
+                    }
+                };
+            }
+        });
+    </script>
+    <script type="module" src="/js/main.js" defer></script>
 </body>
 </html>
